Bubble Shooting-Game
The project Balloon Shooting is a gaming application that uses C language to run the application. This was created with children as the major customers.
This application is mainly intended to provide a gaming application that runs on a desktop environment. The game uses less amount memory along with the necessary graphics to the end-user.
This is a game which consists of three levels. In each level the user is provided with a set of balloons and arrows. The number of arrows and balloons varies with increasing levels.
The user has to shoot the balloons with the given arrows so that the level is cleared and the next level is enabled.
Every level has threshold points which must be satisfies so that the user can proceed to the next level.
All the necessary operations and rules for the game will also be provided to the user before the user begins the game.
Finally, this application is aimed at entertaining any user who plays the game.